Add linting, and prove the supported Python range
The project claimed Python 3.11 with nothing checking either the claim or the
style of the code.

- ruff is configured in code/pyproject.toml and runs clean across the package,
  the experiments and the tests. The Vietnamese-text rules (RUF001-003) and
  __all__ sorting are switched off deliberately; everything else is fixed
  rather than suppressed, including three unguarded zip() calls, a redundant
  int(round(...)) and a redundant int(math.floor(...)).
- The suite was run on Python 3.10, 3.11, 3.12 and 3.13 locally: 86 tests pass
  on each. requires-python and the badge now say 3.10, which is what the code
  actually needs (zip(strict=...)), and CI runs the matrix so the claim stays
  honest.
- CI is now five jobs: lint, tests (matrix), reproduce, matlab, report.
- `make lint` replaces the placeholder compileall target.

The split in the reproduce job is now backed by evidence rather than a guess:
running run_all.py under Python 3.10 leaves Sections/generated/*.tex
byte-identical while every PNG in figures/ changes, so the LaTeX tables are a
hard failure when stale and the figures only a warning.

+++ code/lmm/latex.py
@@ -30,7 +30,7 @@ def sci(x: float, digits: int = 3) -> str:
     """Format a float as ``a.bcd \\times 10^{e}`` for maths mode."""
     if x == 0 or not math.isfinite(x):
         return "0" if x == 0 else r"\infty"
-    exponent = int(math.floor(math.log10(abs(x))))
+    exponent = math.floor(math.log10(abs(x)))
     mantissa = x / 10.0**exponent
     if exponent == 0:
         return f"{mantissa:.{digits}f}"
